+++ egui_demo_lib/src/apps/demo/code_example.rs
@@ -98,9 +98,8 @@ impl CodeExample {
         );
 
         ui.horizontal(|ui| {
-            let indentation = 8.0
-                * ui.fonts()
-                    .glyph_width(&egui::TextStyle::Monospace.resolve(ui.style()), ' ');
+            let font_id = egui::TextStyle::Monospace.resolve(ui.style());
+            let indentation = 8.0 * ui.fonts().glyph_width(&font_id, ' ');
             let item_spacing = ui.spacing_mut().item_spacing;
             ui.add_space(indentation - item_spacing.x);